+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Changelog for vega-view
 
+## 0.2
+
+The directory view now displays all the visualizations in a directory,
+and hides those files that are not JSON. Selecting a visualization can
+now open it in a new page (0.1 behavior) or inline (new).
+
 ## 0.1
 
 Initial release